Clean up tool use types
