--- src/core/tools/types.ts
@@ -1,19 +1,0 @@
-import { ClineAsk, ToolProgressStatus } from "../../schemas"
-import { ToolParamName } from "../assistant-message"
-import { ToolResponse } from "../Cline"
-
-export type AskApproval = (
-	type: ClineAsk,
-	partialMessage?: string,
-	progressStatus?: ToolProgressStatus,
-) => Promise<boolean>
-
-export type HandleError = (action: string, error: Error) => Promise<void>
-
-export type PushToolResult = (content: ToolResponse) => void
-
-export type RemoveClosingTag = (tag: ToolParamName, content?: string) => string
-
-export type AskFinishSubTaskApproval = () => Promise<boolean>
-
-export type ToolDescription = () => string
